feat: populate Blog card with real Ghost posts via Server Component
- Extract HomePage client logic into _components/HomeClient.tsx
- New page.tsx is an async Server Component that fetches latest Ghost posts
  via getLatestPosts(4) and passes them to HomeClient as blogPosts prop
- BlogContent and MobileBlog accept optional posts prop, render real
  post.title, post.excerpt, post.url, and post.primary_tag (color-coded)
- Graceful fallback mock posts if Ghost API is unavailable
- Fixes commit authorship (PorteGumi <port@gitchegumi.com>)
