+++ public/sitemap-0.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://www.gitchegumi.com</loc><lastmod>2026-05-27T03:33:40.669Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.gitchegumi.com/about</loc><lastmod>2026-05-27T03:33:40.669Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.gitchegumi.com/accountpipe</loc><lastmod>2026-05-27T03:33:40.669Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.gitchegumi.com/blog</loc><lastmod>2026-05-27T03:33:40.669Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.gitchegumi.com/budget</loc><lastmod>2026-05-27T03:33:40.669Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.gitchegumi.com/debtpipe</loc><lastmod>2026-05-27T03:33:40.669Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.gitchegumi.com/portfolio</loc><lastmod>2026-05-27T03:33:40.669Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.gitchegumi.com/tools</loc><lastmod>2026-05-27T03:33:40.669Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.gitchegumi.com/tools/debtpipe</loc><lastmod>2026-05-27T03:33:40.669Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.gitchegumi.com/trakpipe</loc><lastmod>2026-05-27T03:33:40.669Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.gitchegumi.com/voice-over</loc><lastmod>2026-05-27T03:33:40.669Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.gitchegumi.com</loc><lastmod>2026-05-27T04:12:40.596Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.gitchegumi.com/about</loc><lastmod>2026-05-27T04:12:40.596Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.gitchegumi.com/accountpipe</loc><lastmod>2026-05-27T04:12:40.596Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.gitchegumi.com/blog</loc><lastmod>2026-05-27T04:12:40.596Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.gitchegumi.com/budget</loc><lastmod>2026-05-27T04:12:40.596Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.gitchegumi.com/debtpipe</loc><lastmod>2026-05-27T04:12:40.596Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.gitchegumi.com/portfolio</loc><lastmod>2026-05-27T04:12:40.596Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.gitchegumi.com/tools</loc><lastmod>2026-05-27T04:12:40.596Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.gitchegumi.com/tools/debtpipe</loc><lastmod>2026-05-27T04:12:40.596Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.gitchegumi.com/trakpipe</loc><lastmod>2026-05-27T04:12:40.596Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.gitchegumi.com/voice-over</loc><lastmod>2026-05-27T04:12:40.596Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
